Isolate playbook types
